+++ src/templates/blogTemplate.tsx
@@ -5,8 +5,8 @@ import styles from "./blogTemplate.module.css"
 import { BlogTemplateQuery } from "../../types/graphql-types"
 import Layout from "../components/layout"
 import SEO from "../components/seo"
-import Toc from "../components/toc"
-import TocMobile from "../components/tocMoblie"
+import Toc from "../components/article/toc/toc"
+import TocMobile from "../components/article/toc/tocMoblie"
 import Social from "../components/article/social/socials"
 import SocialMobile from "../components/article/social/socialsMobile"
 import { Tag } from "../components/tag"
